MOON-272: Use new typescript format for Dropdown component
